api/requirements.txt
aio-pika==6.4.3
aiormq==3.2.0
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.3.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
djangorestframework==3.12.4
drf-yasg==1.20.0
grpcio==1.43.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==2.8
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
messages==0.8.0
multidict==4.7.4
packaging==20.9
pamqp==2.3.0
pika==1.1.0
pika-stubs==0.1.3
protobuf==3.19.4
psycopg2==2.9.3
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rfc3986==1.5.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.26.4
validus==0.3.0
yarl==1.4.2
aio-pika==6.4.3
aiormq==3.2.0
aiosmtplib==1.1.6
anyio==3.5.0
asgiref==3.3.4
attrs==22.1.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
djangorestframework==3.12.4
drf-yasg==1.20.0
grpcio==1.43.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==2.8
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
messages==0.8.0
multidict==4.7.4
packaging==20.9
pamqp==2.3.0
pika==1.1.0
pika-stubs==0.1.3
pluggy==1.0.0
protobuf==3.19.4
psycopg2==2.9.3
py==1.11.0
pyparsing==2.4.7
pytest==7.1.2
pytest-django==4.5.2
pytz==2021.1
requests==2.25.1
rfc3986==1.5.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
tomli==2.0.1
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.26.4
validus==0.3.0
yarl==1.4.2
